Implement groupBy in Range and NumericRange
